requirements.txt
attrs==24.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
filelock==3.15.4
Flask==2.1.2
flatbuffers==24.3.25
fsspec==2024.6.1
humanfriendly==10.0
idna==3.8
imageio==2.35.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.1
onnxruntime==1.19.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
Pillow==9.1.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==5.27.3
PyMatting==1.1.12
referencing==0.35.1
rembg==2.0.55
requests==2.32.3
rpds-py==0.20.0
scikit-image==0.24.0
scipy==1.14.1
setuptools==73.0.1
sympy==1.13.2
tifffile==2024.8.10
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==2.1.2
